Rename wiki.core.wiki_paginate to wiki.core.paginator
