+++ src/wiki/plugins/attachments/views.py
@@ -12,7 +12,7 @@ from django.views.generic.base import TemplateView, View
 from django.views.generic.edit import FormView
 from django.views.generic.list import ListView
 from wiki.core.http import send_file
-from wiki.core.wiki_paginate import WikiPaginator
+from wiki.core.paginator import WikiPaginator
 from wiki.decorators import get_article, response_forbidden
 from wiki.plugins.attachments import forms, models, settings
 from wiki.views.mixins import ArticleMixin
