+++ src/wiki/plugins/images/views.py
@@ -11,7 +11,7 @@ from django.views.generic.base import RedirectView
 from django.views.generic.edit import FormView
 from django.views.generic.list import ListView
 from wiki.conf import settings as wiki_settings
-from wiki.core.wiki_paginate import WikiPaginator
+from wiki.core.paginator import WikiPaginator
 from wiki.decorators import get_article
 from wiki.models.pluginbase import RevisionPluginRevision
 from wiki.plugins.images import forms, models
